refactor: move the group display to the GroupsSection component
- Introduced a new GroupsSection component to manage participant grouping within sessions.
- Implemented functionality for auto, manual, and class-based grouping modes.
- Integrated notifications for group application success and error handling.
- Updated HostView to utilize the new GroupsSection, streamlining the UI and improving user experience.
